Second stage of refactor
